Rename labelImg.py to labelImgPlusPlus.py for consistency
Fixes #51

- Rename main application file to match package name
- Update pyproject.toml entry point and py-modules
- Update CI/CD workflows (pyinstaller commands)
- Update build scripts
- Update test imports
- Update all documentation references
