+++ build-tools/build-ubuntu-binary.sh
@@ -11,7 +11,7 @@ pyinstaller --hidden-import=xml \
             --hidden-import=xml.etree \
             --hidden-import=xml.etree.ElementTree \
             --hidden-import=lxml.etree \
-            -D -F -n labelImgPlusPlus -c "../labelImg.py" -p ../libs -p ../
+            -D -F -n labelImgPlusPlus -c "../labelImgPlusPlus.py" -p ../libs -p ../
 
 FOLDER=$(git describe --abbrev=0 --tags)
 FOLDER="linux_"$FOLDER
