+++ labelimgplusplus.egg-info/SOURCES.txt
@@ -1,0 +1,40 @@
+CONTRIBUTING.rst
+HISTORY.rst
+LICENSE
+MANIFEST.in
+README.rst
+labelImg.py
+pyproject.toml
+resources.qrc
+setup.cfg
+data/predefined_classes.txt
+labelimgplusplus.egg-info/PKG-INFO
+labelimgplusplus.egg-info/SOURCES.txt
+labelimgplusplus.egg-info/dependency_links.txt
+labelimgplusplus.egg-info/entry_points.txt
+labelimgplusplus.egg-info/requires.txt
+labelimgplusplus.egg-info/top_level.txt
+libs/__init__.py
+libs/canvas.py
+libs/colorDialog.py
+libs/combobox.py
+libs/commands.py
+libs/constants.py
+libs/create_ml_io.py
+libs/default_label_combobox.py
+libs/galleryWidget.py
+libs/hashableQListWidgetItem.py
+libs/labelDialog.py
+libs/labelFile.py
+libs/lightWidget.py
+libs/pascal_voc_io.py
+libs/resources.py
+libs/settings.py
+libs/shape.py
+libs/stringBundle.py
+libs/styles.py
+libs/toolBar.py
+libs/ustr.py
+libs/utils.py
+libs/yolo_io.py
+libs/zoomWidget.py
