+++ labelimgplusplus.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+labelImgPlusPlus = labelImg:main
