+++ labelimgplusplus.egg-info/requires.txt
@@ -1,0 +1,2 @@
+pyqt5
+lxml
