+++ labelimgplusplus.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+labelImg
+libs
