+++ sonar-project.properties
@@ -8,7 +8,7 @@ sonar.projectName=labelImg++
 sonar.projectVersion=2.0.0
 
 # Source code location
-sonar.sources=labelImg.py,libs
+sonar.sources=labelImgPlusPlus.py,libs
 sonar.tests=tests
 
 # Python specific settings
